Fix: move Header component and its styles to a separated folder
