+++ front/src/pages/App/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import Header from './Header';
+import Header from './Header/Header';
 import Footer from './Footer';
 import Home from '../Home';
 import AddItem from '../AddItem';
